feat: EarthQuakes as functional component & ts
